Moving WafFeaturedCard to own file
