+++ src/views/homepage/components/featured-content-grid/featured-content-grid.module.css
@@ -31,28 +31,3 @@
 .gridAreaD {
 	grid-area: d;
 }
-
-/*
-***
-* WAF FeaturedCard
-***
-*/
-
-.wafCard {
-	border: none;
-	padding-top: 0;
-	width: 373px;
-}
-
-/* TODO replace magic numbers with variables */
-.wafCardContent {
-	margin-bottom: 48px;
-	padding-top: calc(179px - 25px + 30px);
-}
-
-.wafGraphic svg {
-	left: 0;
-	position: absolute;
-	right: 0;
-	top: -25px;
-}
