+++ src/views/homepage/components/featured-content-grid/index.tsx
@@ -3,38 +3,11 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-// HashiCorp imports
-import InlineSvg from '@hashicorp/react-inline-svg'
-
-// Global imports
-import Heading from 'components/heading'
-
-// Local imports
-import wafGraphic from '../../img/waf-graphic.svg?include'
 import CertificationsFeaturedCard from './certifications-featured-card'
-import FeaturedCard from './featured-card'
 import HcpFeaturedCard from './hcp-featured-card'
 import SearchFeaturedCard from './search-featured-card'
+import WafFeaturedCard from './waf-featured-card'
 import s from './featured-content-grid.module.css'
-
-const WafFeaturedCard = () => {
-	const title = 'What is HashiCorp’s Well-Architected Framework?'
-
-	return (
-		<FeaturedCard
-			className={s.wafCard}
-			href="/well-architected-framework"
-			title={title}
-		>
-			<div className={s.wafCardContent}>
-				<InlineSvg className={s.wafGraphic} src={wafGraphic} />
-				<Heading level={2} size={400} weight="bold">
-					{title}
-				</Heading>
-			</div>
-		</FeaturedCard>
-	)
-}
 
 const FeaturedContentGrid = () => {
 	return (
